[feat/#2]-키오스크 컴포넌트 분리
