+++ meal-ticket-system-main/src/styles/kioskMenuPage.css
@@ -15,11 +15,20 @@
   background: #007bff;
   padding: 10px 20px;
   display: flex;
-  justify-content: flex-start;
+  justify-content: space-between;
   align-items: center;
   box-shadow: 0 2px 10px rgba(0,123,255,0.2);
   border-radius: 15px 15px 0 0;
   margin: -30px -20px 20px -20px;
+}
+
+.restaurant-name h1 {
+  color: #fff;
+  font-size: 1.5rem;
+  font-weight: 700;
+  margin: 0;
+  text-align: center;
+  flex: 1;
 }
 
 .kiosk-back-btn {
@@ -320,6 +329,12 @@
   .kiosk-header {
     padding: 8px 15px;
     margin: -20px -15px 15px -15px;
+    flex-direction: column;
+    gap: 8px;
+  }
+
+  .restaurant-name h1 {
+    font-size: 1.2rem;
   }
   
   .kiosk-back-btn {
@@ -419,6 +434,12 @@
   .kiosk-header {
     padding: 6px 10px;
     margin: -15px -10px 10px -10px;
+    flex-direction: column;
+    gap: 6px;
+  }
+
+  .restaurant-name h1 {
+    font-size: 1rem;
   }
   
   .kiosk-back-btn {
